+++ backbone.dualstorage.js
@@ -25,14 +25,14 @@ as that.
     return _.result(collection, 'storeName') || _.result(model, 'storeName') || _.result(collection, 'url') || _.result(model, 'urlRoot') || _.result(model, 'url');
   };
 
-  Backbone.Collection.prototype.syncDirty = function() {
+  Backbone.Collection.prototype.syncDirty = function(options) {
     var id, ids, store, _i, _len, _ref, _results;
     store = localStorage.getItem("" + (getStoreName(this)) + "_dirty");
     ids = (store && store.split(',')) || [];
     _results = [];
     for (_i = 0, _len = ids.length; _i < _len; _i++) {
       id = ids[_i];
-      _results.push((_ref = this.get(id)) != null ? _ref.save() : void 0);
+      _results.push((_ref = this.get(id)) != null ? _ref.save(null, options) : void 0);
     }
     return _results;
   };
@@ -53,7 +53,7 @@ as that.
     return _.compact(models);
   };
 
-  Backbone.Collection.prototype.syncDestroyed = function() {
+  Backbone.Collection.prototype.syncDestroyed = function(options) {
     var id, ids, model, store, _i, _len, _results;
     store = localStorage.getItem("" + (getStoreName(this)) + "_destroyed");
     ids = (store && store.split(',')) || [];
@@ -63,7 +63,7 @@ as that.
       model = new this.model;
       model.set(model.idAttribute, id);
       model.collection = this;
-      _results.push(model.destroy());
+      _results.push(model.destroy(options));
     }
     return _results;
   };
@@ -74,9 +74,9 @@ as that.
     return ids = (store && store.split(',')) || [];
   };
 
-  Backbone.Collection.prototype.syncDirtyAndDestroyed = function() {
-    this.syncDirty();
-    return this.syncDestroyed();
+  Backbone.Collection.prototype.syncDirtyAndDestroyed = function(options) {
+    this.syncDirty(options);
+    return this.syncDestroyed(options);
   };
 
   S4 = function() {
@@ -367,9 +367,9 @@ as that.
         if (localsync('hasDirtyOrDestroyed', model, options)) {
           return useOfflineStorage();
         } else {
-          options.success = function(resp, status, xhr) {
+          options.success = function(resp, _status, _xhr) {
             var collection, idAttribute, modelAttributes, responseModel, _i, _len;
-            if (hasOfflineStatusCode(xhr)) {
+            if (hasOfflineStatusCode(options.xhr)) {
               return useOfflineStorage();
             }
             resp = parseRemoteResponse(model, resp);
@@ -393,34 +393,34 @@ as that.
               responseModel = modelUpdatedWithResponse(model, resp);
               localsync('update', responseModel, options);
             }
-            return success(resp, status, xhr);
+            return success(resp, _status, _xhr);
           };
           options.error = function(xhr) {
             return relayErrorCallback(xhr);
           };
-          return onlineSync(method, model, options);
+          return options.xhr = onlineSync(method, model, options);
         }
         break;
       case 'create':
-        options.success = function(resp, status, xhr) {
+        options.success = function(resp, _status, _xhr) {
           var updatedModel;
-          if (hasOfflineStatusCode(xhr)) {
+          if (hasOfflineStatusCode(options.xhr)) {
             return useOfflineStorage();
           }
           updatedModel = modelUpdatedWithResponse(model, resp);
           localsync(method, updatedModel, options);
-          return success(resp, status, xhr);
+          return success(resp, _status, _xhr);
         };
         options.error = function(xhr) {
           return relayErrorCallback(xhr);
         };
-        return onlineSync(method, model, options);
+        return options.xhr = onlineSync(method, model, options);
       case 'update':
         if (model.hasTempId()) {
           temporaryId = model.id;
-          options.success = function(resp, status, xhr) {
+          options.success = function(resp, _status, _xhr) {
             var updatedModel;
-            if (hasOfflineStatusCode(xhr)) {
+            if (hasOfflineStatusCode(options.xhr)) {
               return useOfflineStorage();
             }
             updatedModel = modelUpdatedWithResponse(model, resp);
@@ -429,7 +429,7 @@ as that.
             });
             localsync('delete', model, options);
             localsync('create', updatedModel, options);
-            return success(resp, status, xhr);
+            return success(resp, _status, _xhr);
           };
           options.error = function(xhr) {
             model.set(model.idAttribute, temporaryId, {
@@ -440,21 +440,21 @@ as that.
           model.set(model.idAttribute, null, {
             silent: true
           });
-          return onlineSync('create', model, options);
+          return options.xhr = onlineSync('create', model, options);
         } else {
-          options.success = function(resp, status, xhr) {
+          options.success = function(resp, _status, _xhr) {
             var updatedModel;
-            if (hasOfflineStatusCode(xhr)) {
+            if (hasOfflineStatusCode(options.xhr)) {
               return useOfflineStorage();
             }
             updatedModel = modelUpdatedWithResponse(model, resp);
             localsync(method, updatedModel, options);
-            return success(resp, status, xhr);
+            return success(resp, _status, _xhr);
           };
           options.error = function(xhr) {
             return relayErrorCallback(xhr);
           };
-          return onlineSync(method, model, options);
+          return options.xhr = onlineSync(method, model, options);
         }
         break;
       case 'delete':
@@ -462,17 +462,17 @@ as that.
           options.ignoreCallbacks = false;
           return localsync(method, model, options);
         } else {
-          options.success = function(resp, status, xhr) {
-            if (hasOfflineStatusCode(xhr)) {
+          options.success = function(resp, _status, _xhr) {
+            if (hasOfflineStatusCode(options.xhr)) {
               return useOfflineStorage();
             }
             localsync(method, model, options);
-            return success(resp, status, xhr);
+            return success(resp, _status, _xhr);
           };
           options.error = function(xhr) {
             return relayErrorCallback(xhr);
           };
-          return onlineSync(method, model, options);
+          return options.xhr = onlineSync(method, model, options);
         }
     }
   };