+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 * Fix #115: Don't attempt to delete models from the server that were only local, and call the callback in this case (Micha Reiser)
 * Allow Backbone.DualStorage.offlineStatusCodes to include 200 OK as an offline status
+* syncDirty, syncDestroyed, and syncDirtyAndDestroyed accept options (and therefore callbacks) which are passed to save and destroy
 
 ### 1.3.1 / 2014-15-30
 
@@ -16,6 +17,7 @@
 * Fix #105: Always set options.dirty in callbacks when offline (Elad Efrat)
 * Fix #93: syncDirty works when models use a custom idAttribute (Ben Salinas)
 * Fix #78: Do not clear local collection cache when a model is fetched
+* Add Backbone.DualStorage.offlineStatusCodes for configuring what to consider as offline
 * Add `make watch` for continual coffeescript compilation during development
 * `make` compiles sourcemaps between coffeescript and javascript
 * Prevent id duplication in the internal list of model ids for a collection when a model is fetched